feat: only create subscription if no subscription is currently active

+++ subscribe.php
@@ -13,6 +13,13 @@ if (isset($token) && isset($email) && isset($stripePriceId)) {
     try {
         $user = findRecord('users', ['email' => $email]);
         $product = findRecord('products', ['stripe_price_id' => $stripePriceId]);
+        $existingSubscription = findRecord(
+            'subscriptions',
+            [
+                'user_id' => $user['id'],
+                'status' => 'active'
+            ]
+        );
 
         $stripeCustomerId = $user['stripe_customer_id'];
 
@@ -32,31 +39,36 @@ if (isset($token) && isset($email) && isset($stripePriceId)) {
             );
         }
 
-        // Create a new subscription
-        $subscription = \Stripe\Subscription::create([
-            'customer' => $stripeCustomerId,
-            'items' => [
-                [
-                    'price' => $stripePriceId,
+        // Only create subscription if no subscription is currently active
+        if (!isset($existingSubscription)) {
+            // Create a new subscription
+            $subscription = \Stripe\Subscription::create([
+                'customer' => $stripeCustomerId,
+                'items' => [
+                    [
+                        'price' => $stripePriceId,
+                    ],
                 ],
-            ],
-        ]);
+            ]);
 
-        // Save the subscription in the DB
-        $subscriptionRecord = createRecord(
-            'subscriptions',
-            [
-                'user_id' => $user['id'],
-                'product_id' => $product['id'],
-                'stripe_subscription_id' => $subscription->id,
-                'stripe_price_id' => $stripePriceId,
-                'status' => 'active',
-                'start_date' =>  date('Y-m-d H:i:s', $subscription->current_period_start),
-                'end_date' => date('Y-m-d H:i:s', $subscription->current_period_start)
-            ]
-        );
+            // Save the subscription in the DB
+            $subscriptionRecord = createRecord(
+                'subscriptions',
+                [
+                    'user_id' => $user['id'],
+                    'product_id' => $product['id'],
+                    'stripe_subscription_id' => $subscription->id,
+                    'stripe_price_id' => $stripePriceId,
+                    'status' => 'active',
+                    'start_date' =>  date('Y-m-d H:i:s', $subscription->current_period_start),
+                    'end_date' => date('Y-m-d H:i:s', $subscription->current_period_start)
+                ]
+            );
 
-        echo json_encode(['success' => true, 'subscription' => $subscription]);
+            echo json_encode(['success' => true, 'subscription' => $subscription]);
+        } else {
+            echo json_encode(['success' => false, 'message' => "Can't subscribe to the plan. You already have active subscription."]);
+        }
     } catch (\Stripe\Exception\ApiErrorException $e) {
         echo json_encode(['error' => $e->getMessage()]);
     }
